Extract Aspire.Hosting.AWS.Tests
