+++ src/Aspire.Hosting.AWS/Aspire.Hosting.AWS.csproj
@@ -26,7 +26,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <InternalsVisibleTo Include="Aspire.Hosting.Tests" />
+    <InternalsVisibleTo Include="Aspire.Hosting.AWS.Tests" />
   </ItemGroup>
 
 </Project>
